FEAT: window

+++ src/Components/Navbar.jsx
@@ -1,13 +1,19 @@
 import { lang } from "../App";
 import "../scss/navbar.scss"
 
-export default function Navbar() {
+export default function Navbar({ refs }) {
     const txt = lang[localStorage.getItem("lang")];
 
     function changeLanguage() {
-        if (localStorage.getItem("lang") === "en") { localStorage.setItem("lang", "fr") }
-        else { localStorage.setItem("lang", "en") }
-        window.location.reload();
+        if (confirm(txt["warning_translate"])) {
+            if (localStorage.getItem("lang") === "en") { localStorage.setItem("lang", "fr") }
+            else { localStorage.setItem("lang", "en") }
+            window.location.reload();
+        }
+    }
+
+    function openWindow(ref) {
+        ref.current.style.display = "block"
     }
 
     return (
@@ -19,6 +25,7 @@ export default function Navbar() {
                     <button className="WIP"><i className='bx bx-save'></i> {txt.save}</button>
                     <button className="WIP"><i className='bx bx-import'></i> {txt.import}</button>
                     <button className="WIP"><i className='bx bx-export'></i> {txt.export}</button>
+                    <button className="WIP"><i className='bx bx-cloud'></i> {txt.browser_projects}</button>
                     <button className="WIP"><i className='bx bxs-palette'></i> {txt.theme}</button>
                     <button onClick={changeLanguage}><i className='bx bxs-flag-alt'></i> {txt.language}</button>
                 </div>
@@ -26,9 +33,9 @@ export default function Navbar() {
             <article>
                 <div className="navTitle">{txt.project}<i className='bx bx-chevron-down'></i></div>
                 <div className="dropdown">
-                    <button className="WIP">{txt.general}</button>
-                    <button className="WIP">{txt.startMenu}</button>
-                    <button className="WIP">{txt.display}</button>
+                    <button onClick={() => { openWindow(refs.general) }} >{txt.general}</button>
+                    <button onClick={() => { openWindow(refs.startMenu) }} >{txt.startMenu}</button>
+                    <button onClick={() => { openWindow(refs.display) }} >{txt.display}</button>
                     <button className="WIP">{txt.style}</button>
                 </div>
             </article>
